stock-news-scraper/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
Automat==25.4.16
bcrypt==4.0.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
constantly==23.10.4
cryptography==44.0.3
cssselect==1.3.0
defusedxml==0.7.1
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
Faker==37.4.0
fastapi==0.115.12
fastapi-cli==0.0.7
filelock==3.18.0
greenlet==3.1.1
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itemadapter==0.11.0
itemloaders==1.3.2
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
lxml==6.0.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mysql-connector-python==9.3.0
orjson==3.10.18
packaging==25.0
parsel==1.10.0
passlib==1.7.4
playwright==1.51.0
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.22.1
Protego==0.5.0
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.4
pydantic-extra-types==2.10.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
PyDispatcher==2.0.7
pyee==12.1.1
Pygments==2.19.1
pyOpenSSL==25.1.0
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
PyYAML==6.0.2
queuelib==1.8.0
requests==2.32.3
requests-file==2.1.0
rich==14.0.0
rich-toolkit==0.14.5
rsa==4.9.1
Scrapy==2.13.3
service-identity==24.2.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
starlette_exporter==0.23.0
tldextract==5.3.0
Twisted==25.5.0
typer==0.15.3
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.2
ujson==5.10.0
urllib3==2.4.0
uvicorn==0.34.2
w3lib==2.3.1
watchfiles==1.0.5
websockets==15.0.1
zope.interface==7.2
